Pause on fragment change and display option in agenda
