+++ app/src/main/java/com/orgzly/android/ui/fragments/QueryFragment.java
@@ -1,5 +1,6 @@
 package com.orgzly.android.ui.fragments;
 
+import android.app.AlertDialog;
 import android.content.Context;
 import android.database.Cursor;
 import android.os.Bundle;
@@ -10,6 +11,8 @@ import android.support.v4.widget.SimpleCursorAdapter;
 import android.support.v7.view.ActionMode;
 import android.view.Menu;
 import android.view.MenuInflater;
+import android.view.MenuItem;
+import android.view.WindowManager;
 
 import com.orgzly.BuildConfig;
 import com.orgzly.R;
@@ -140,6 +143,58 @@ abstract public class QueryFragment extends NoteListFragment
 
         mListAdapter.changeCursor(null);
     }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        if (BuildConfig.LOG_DEBUG) LogUtils.d(TAG, savedInstanceState);
+        super.onCreate(savedInstanceState);
+
+        setHasOptionsMenu(true);
+    }
+
+    /*
+     * Options menu.
+     */
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        if (BuildConfig.LOG_DEBUG) LogUtils.d(TAG, menu, inflater);
+
+        inflater.inflate(R.menu.query_actions, menu);
+        boolean keepScreenOnEnabled = (getActivity().getWindow().getAttributes().flags & WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON) != 0;
+        menu.findItem(R.id.query_options_keep_screen_on).setChecked(keepScreenOnEnabled);
+        super.onCreateOptionsMenu(menu, inflater);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (BuildConfig.LOG_DEBUG) LogUtils.d(TAG, item);
+
+        switch (item.getItemId()) {
+            case R.id.query_options_keep_screen_on:
+                boolean keepScreenOnEnabled = (getActivity().getWindow().getAttributes().flags & WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON) != 0;
+                if (!keepScreenOnEnabled) {
+                    dialog = new AlertDialog.Builder(getActivity())
+                            .setTitle(R.string.keep_screen_on)
+                            .setMessage(R.string.keep_screen_on_desc)
+                            .setPositiveButton(android.R.string.yes, (dialog, which) -> {
+                                getActivity().getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+                                item.setChecked(true);
+                                dialog.dismiss();
+                            })
+                            .setNegativeButton(android.R.string.cancel, (dialog, id) -> dialog.dismiss())
+                            .create();
+                    dialog.show();
+                } else {
+                    getActivity().getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+                    item.setChecked(false);
+                }
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
+
 
     @Override
     abstract public ActionMode.Callback getNewActionMode();
