+++ app/src/main/res/menu/query_actions.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/search_options_keep_screen_on"
+        android:id="@+id/query_options_keep_screen_on"
         app:showAsAction="never"
         android:checkable="true"
         android:title="@string/keep_screen_on"/>
